+++ stereotomy/src/main/java/com/salesforce/apollo/stereotomy/event/DelegatingEventCoordinates.java
@@ -1,5 +1,7 @@
 package com.salesforce.apollo.stereotomy.event;
 
+import org.joou.ULong;
+
 import com.salesfoce.apollo.stereotomy.event.proto.DelegatingEventCoords;
 import com.salesforce.apollo.stereotomy.EventCoordinates;
 import com.salesforce.apollo.stereotomy.identifier.Identifier;
@@ -9,17 +11,17 @@ public class DelegatingEventCoordinates {
     private final Identifier       identifier;
     private final String           ilk;
     private final EventCoordinates previousEvent;
-    private final long             sequenceNumber;
+    private final ULong            sequenceNumber;
 
     public DelegatingEventCoordinates(DelegatingEventCoords coordinates) {
         identifier = Identifier.from(coordinates.getIdentifier());
-        sequenceNumber = coordinates.getSequenceNumber();
+        sequenceNumber = ULong.valueOf(coordinates.getSequenceNumber());
         previousEvent = new EventCoordinates(coordinates.getPrevious());
         ilk = coordinates.getIlk();
     }
 
-    public DelegatingEventCoordinates(String ilk, Identifier identifier, long sequenceNumber,
-            EventCoordinates previousEvent) {
+    public DelegatingEventCoordinates(String ilk, Identifier identifier, ULong sequenceNumber,
+                                      EventCoordinates previousEvent) {
         this.identifier = identifier;
         this.sequenceNumber = sequenceNumber;
         this.previousEvent = previousEvent;
@@ -38,14 +40,14 @@ public class DelegatingEventCoordinates {
         return previousEvent;
     }
 
-    public long getSequenceNumber() {
+    public ULong getSequenceNumber() {
         return sequenceNumber;
     }
 
     public DelegatingEventCoords toCoords() {
         return DelegatingEventCoords.newBuilder()
                                     .setIdentifier(identifier.toIdent())
-                                    .setSequenceNumber(sequenceNumber)
+                                    .setSequenceNumber(sequenceNumber.longValue())
                                     .setPrevious(previousEvent.toEventCoords())
                                     .setIlk(ilk)
                                     .build();
